+++ apps/backend/src/user-management-app/auth/interfaces/jwt-payload.interface.ts
@@ -1,0 +1,9 @@
+/**
+ * Represents the payload for the JWT strategy.
+ */
+export interface JwtPayload {
+  userId: number;
+  mfaVerified: boolean;
+  iat?: number;
+  exp?: number;
+}
