+++ apps/backend/src/user-management-app/auth/strategies/jwt.strategy.ts
@@ -6,13 +6,7 @@ import { UsersService } from 'src/user-management-app/users/users.service';
 import { plainToInstance } from 'class-transformer';
 import { UserEntity } from '../../users/entities/user.entity';
 import { Request } from 'express';
-
-export interface JwtPayload {
-  userId: number;
-  mfaVerified: boolean;
-  iat?: number;
-  exp?: number;
-}
+import { JwtPayload } from '../interfaces/jwt-payload.interface';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy, 'jwt') {
